backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic-settings==2.0.3
alembic==1.12.1
python-dotenv==1.0.0